requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.2
django-bootstrap5==21.3
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-gmailapi-backend==0.3.2
django-timedeltatemplatefilter==0.1.2
django-widget-tweaks==1.4.12
environ==1.0
et-xmlfile==1.1.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
Japanera==0.1.1
kanjize==1.0.0
numpy==1.23.0
openpyxl==3.0.9
pdfrw==0.4
Pillow==9.1.1
protobuf==4.21.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyPDF2==2.4.0
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.7
requests==2.28.0
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
